chatbot/requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit
langchain-community
lanserver
fastapi
uvicorn